Add a preprocess step that injects inset text
Redcarpet offers a preprocess[0] step that allows us to transform the
document either before or after the Markdown rendering. As HTML is valid
markdown, it's safe to generate and inject the HTML before the regular
Redcarpet rendering takes place.

The implementation here covers inset text[1] but should work for styles
like warning text[2] or tag[3].

The suggested markup here is:

```
{inset-text}Text to be inset goes here{/inset-text}
```

Alternatively, to split across multiple lines:

```
{inset-text}
  Text to be
  inset goes
  here
{/inset-text}
```

[0] https://github.com/vmg/redcarpet#prepost-process
[1] https://design-system.service.gov.uk/components/inset-text/
[2] https://design-system.service.gov.uk/components/warning-text/
[3] https://design-system.service.gov.uk/components/tag/