Add front matter stripping option to renderer
Sometimes front matter is included in files but we don't want it to be
part of the final document. Markdown fences are intepreted as horizontal
rules and the keys and values within are rendered in paragraphs, which
looks ugly.

The new strip_front_matter argument, which defaults to true, removes
three dashes at the beginning of the document up to a matching three
dash closing fence.